Hide redundant move-to-top actions when episode already leading queue (#73)
* Hide redundant move-to-top actions when already at queue start

* ✨ refactor(queue): simplify queue order checks

This commit refactors the logic for checking if certain episodes in the queue are not at the top.

Previously, the `anySelectedQueuedNotAtTopOfQueue` property checked if an episode was queued and then if its `queueOrder` was greater than 0. This has been simplified to `anySelectedNotAtTopOfQueue` which directly checks if the `queueOrder` is greater than 0, as the context of being queued is already implied by the presence of `queueOrder`. This reduces redundant checks and improves code readability. The changes are reflected in `SelectableEpisodesToolbarItems.swift`, `EpisodeContextMenuViewModifier.swift`, `EpisodeSwipeViewModifier.swift`, and `SelectableEpisodeList.swift`.

* ✨ refactor(episode-list): simplify queue order check

Simplify the `anySelectedNotAtTopOfQueue` computed property in the `SelectableEpisodeList` protocol. The previous implementation involved checking for an optional `queueOrder` and returning `true` if it was `nil`. This was problematic as a `nil` `queueOrder` should not necessarily be considered "not at the top of the queue".

The updated implementation directly checks if `episode.queueOrder` is not equal to `0`. This correctly identifies episodes that are not at the very beginning of the queue.

* ✨ refactor(ui): add more checks for selected episodes

Add `anySelectedNotAtTopOfQueue` boolean property to `SelectableEpisodeList` protocol extension. This new property checks if any of the selected episodes are not at the top of the queue. This enhances the ability to conditionally enable or disable UI elements based on the queue status of multiple selected episodes. The existing `anySelectedNotAtTopOfQueue` property was also moved to improve code organization.